Changed Scarecrow class name to TargetDummy and added code to ItemFactory to be able to create TargetDummy entity.
