+++ source/core/src/main/com/csse3200/game/entities/factories/ItemFactory.java
@@ -55,6 +55,7 @@ public class ItemFactory {
             case "armor" -> createArmor();
             case "damagebuff" -> createDamageBuff();
             case "reroll" -> createReroll();
+            case "targetdummy" -> createTargetDummy();
             default -> throw new IllegalArgumentException("Invalid item specification: " + specification);
         };
     }
@@ -110,6 +111,10 @@ public class ItemFactory {
     private Collectible createArmor() { return new Armor();}
 
     private Collectible createDamageBuff() { return new DamageBuff();}
+
+    private Collectible createTargetDummy() {
+        return new TargetDummy();
+    }
 
 }
